LayoutTests/platform/blackberry/fast/forms/get-forms-to-about-blank-expected.txt

FAIL: Timed out waiting for notifyDone to be called
Submitting a GET form to <about:blank> should redirect the browser to <about:blank?>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe.contentDocument.URL is "about:blank"
